SaveCopyTask: transfer non-argb888 bitmap tp argb888 before apply filter
BZ: 76653

When GIF image file was edited and saved, the default bitmap type was
kIndex8_Config, so java vm will alloc width*height bytes to handle it.
But when apply filter, the filter by default use argb888 bitmap to
access the allocated buffer, this will corrupt the vm memory chunk.

So we transfer the non-argb888 bitmap to argb888 bitmap before apply any
fileter.

Change-Id: I74619055d2d3338d9c3971c9a24b7dea684fe936
Signed-off-by: wang, biao <biao.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/85470
Reviewed-by: Zhang, Hongyu <hongyu.zhang@intel.com>
Reviewed-by: Shi, Yang A <yang.a.shi@intel.com>
Reviewed-by: Li, XiaojingX <xiaojingx.li@intel.com>
Tested-by: Li, XiaojingX <xiaojingx.li@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>