[PORT FROM MAIN]Video: set the background to clear ghosting images when ActionBar slides in.
BZ: 76841

We need to set the background to clear ghosting
images when ActionBar slides in. However, if we keep the background,
there will be one additional layer in HW composer, which is bad to battery.
As a solution, we remove the background when we hide the action bar.

Change-Id: I1f3cbac06115c73b21569ab569348e03fe4494a9
Orig-Change-Id: I38caf3a85946354698d5e2c65d18fd30a0213905
Category: feature differentiation
Domain: Video.Rendering - HDMI
Origin: internal
Upstream-Candidate: no
Signed-off-by: Zhu,Tianyang <tianyang.zhu@intel.com>
Signed-off-by: Lin Xie <lin.xie@intel.com>