[PORT FROM MAIN]Set the trimVideo orientation to landscape if connected with HDMI.
BZ: 90825

The design of trimVideo has some conflict with our UC of extended mode.
It will reset the whole playback pipeline when it detect the orientation
change.
Screen will be forced to landscape under the extended mode while it
resumes when back to clone mode. So we will see:

 startPlay --> extended mode(Set orientation to landscape) -->
 stopPlay(Detect orientation change) --> clone mode --> restartPlay -->
 extended mode --> stopPlay --> clone mode ......

The fix is to change the screen orientation to landscape if a HDMI
display device is connected.

Category: feature differentiation
Domain: Video.Rendering - HDMI
Origin: internal
Upstream-Candidate: no

Change-Id: I5f5f9146f92ba8a228fd776ee38e06cde6f5a01b
Orig-Change-Id: I81656872890e9d6888fda74fd2e094fe41fc7777
Signed-off-by: Lin Xie <lin.xie@intel.com>