Set the trimVideo orientation to landscape if connected with HDMI.
BZ: 90825

The design of trimVideo has some conflict with our UC of extended mode.
It will reset the whole playback pipeline when it detect the orientation
change.
Screen will be forced to landscape under the extended mode while it
resumes when back to clone mode. So we will see:

 startPlay --> extended mode(Set orientation to landscape) -->
 stopPlay(Detect orientation change) --> clone mode --> restartPlay -->
 extended mode --> stopPlay --> clone mode ......

The fix is to change the screen orientation to landscape if a HDMI
display device is connected.

Change-Id: I81656872890e9d6888fda74fd2e094fe41fc7777
Signed-off-by: Lin Xie <lin.xie@intel.com>
Reviewed-on: http://android.intel.com:8080/95928
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Liang, Dan <dan.liang@intel.com>
Reviewed-by: Chen, Tianmi <tianmi.chen@intel.com>
Reviewed-by: Hu, Jason <jason.hu@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>