Gallery3d: Fix the size is not updated timely
BZ: 109188

Since the mSize is the last album size, if last image is
viewed, index is equal to mSize, so index could be minused
by one, then the last two image could be showed when screen
shot image is stored.

Solution: remove the minus index, and update the items array
to the current size.

Change-Id: I9b1dbdd1e17f33dbadaa4204f06fed5a00b57210
Signed-off-by: jwan <jia.wan@intel.com>
Reviewed-on: http://android.intel.com:8080/108441
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Yao, Yong <yong.yao@intel.com>
Tested-by: Chen, HuaqiangX <huaqiangx.chen@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>