set the background to clear ghosting images when ActionBar slides in
BZ: 76841

We need to set the background to clear ghosting
images when ActionBar slides in. However, if we keep the background,
there will be one additional layer in HW composer, which is bad to battery.
As a solution, we remove the background when we hide the action bar.

Feature: AOSP-Gallery
Category: aosp improvement
Severity: medium
Origin: Internal
Upstream-Candidate: Yes
Orig-Change-Id: I38caf3a85946354698d5e2c65d18fd30a0213905
Change-Id: Ia3cf28d4a0c86d2d1059330b952a9f0e02ce3b5c
Signed-off-by: Zhu,Tianyang <tianyang.zhu@intel.com>
Reviewed-on: http://android.intel.com:8080/89714
Reviewed-by: Xie, Lin <lin.xie@intel.com>
Reviewed-by: Hu, Jason <jason.hu@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: Xu, Randy <randy.xu@intel.com>
Tested-by: Xu, Randy <randy.xu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>