Android Default Camera: disable camera dual open debugging
BZ: 144868

[Description]
Assert() is hit if dual open is detected and the debug flag is being
set to true.

[Implemented Solution]
Disable the DEBUG_OPEN_RELEASE flag and do a error print instead.
Also, the Assert() is relocated so that it will be only hit when
DEBUG_OPEN_RELEASE is enabled, which can be done in development.

Category: feature differentiation
Domain: Camera-Gallery2
Origin: internal
Upstream-Candidate: No, Not AOSP

Change-Id: I696bd2508bb88a5b9dc4a046c51f443b820a90a0
Signed-off-by: Rukai Wei <rukaix.wei@intel.com>
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>